requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
python-dotenv==1.0.0
flake8==6.1.0